models_v2/pytorch/fbnet/inference/gpu/requirements.txt
argparse
numpy
onnx
scipy
timm
torchsummary
